% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonsca_functions.R
\name{inertia_pct}
\alias{inertia_pct}
\title{Percent Inertia from SONSCA}
\usage{
inertia_pct(tab)
}
\arguments{
\item{tab}{A numeric matrix or table.}
}
\value{
Numeric vector of percent inertia values (summing to 100).
}
\description{
Computes the percentage of total inertia explained by each dimension in
a SONSCA solution, using a direct SVD of the standardised residual matrix.
}
